:art: Split up ZGW tasks in submodules
