+++ src/bptl/activiti/tests/test_full_process.py
@@ -41,7 +41,7 @@ class WorkUnitTestCase(TokenAuthMixin, APITestCase):
     def test_full_process(self, m):
         mapping = TaskMapping.objects.create(
             topic_name="zaak-initialize",
-            callback="bptl.work_units.zgw.tasks.CreateZaakTask",
+            callback="bptl.work_units.zgw.tasks.zaak.CreateZaakTask",
         )
         DefaultServiceFactory.create(
             task_mapping=mapping,
