+++ src/bptl/work_units/zgw/tasks/__init__.py
@@ -1,0 +1,4 @@
+from .resultaat import CreateResultaatTask  # noqa
+from .status import CreateStatusTask  # noqa
+from .zaak import CloseZaakTask, CreateZaakTask  # noqa
+from .zaak_relations import RelateDocumentToZaakTask, RelatePand  # noqa
