+++ src/bptl/work_units/zgw/tests/test_get_client.py
@@ -10,7 +10,7 @@ from bptl.tasks.tests.factories import TaskMappingFactory
 from bptl.work_units.zgw.tests.factories import DefaultServiceFactory
 
 from ..client import MultipleServices, NoAuth, NoService
-from ..tasks import ZGWWorkUnit
+from ..tasks.base import ZGWWorkUnit
 
 ZRC_URL = "https://some.zrc.nl/api/v1/"
